FAQViewController, refactoring to VIPER
