+++ VolleyBolley/Modules/Profile/Router/ProfileRouter.swift
@@ -10,6 +10,7 @@ import UIKit
 protocol ProfileRouterProtocol: AnyObject {
     func attachViewController(_ view: UIViewController)
     func showPersonalData()
+    func showFAQ()
 }
 
 final class ProfileRouter: ProfileRouterProtocol {
@@ -26,5 +27,13 @@ final class ProfileRouter: ProfileRouterProtocol {
         }
 
         viewController?.navigationController?.pushViewController(personalDataVC, animated: true)
+    }
+
+    func showFAQ() {
+        guard let faqVC = DIContainer.shared.resolver.resolve(FAQViewController.self) else {
+            fatalError("FAQViewController не зарегистрирован")
+        }
+
+        viewController?.navigationController?.pushViewController(faqVC, animated: true)
     }
 }
